Address all comments
